refactor: Phase 8 — UI decomposition + integration tests (8.1-8.4)
8.1 Settings.tsx: 1,377 → 293 lines (79% reduction). Extracted 11
section components to packages/web/src/components/settings/ with
typed props and self-contained local state.

8.2 System.tsx: 1,352 → 292 lines (78% reduction). Extracted 5 tab
components + overview strip + shared helpers to
packages/web/src/components/system/.

8.3 Search & entity integration tests: already existed (16 search +
24 entity tests). Verified and marked complete.

8.4 MCP tools integration tests: 30 new tests across all 8 MCP tools
(search_brain, list_captures, brain_stats, capture_thought,
get_entity, list_entities, get_weekly_brief, get_capture).

2,585 tests passing, 0 failures.
