+++ packages/web/src/components/system/index.ts
@@ -1,0 +1,6 @@
+export { QueuesTab } from './QueuesTab';
+export { SkillsTab } from './SkillsTab';
+export { FlowsTab } from './FlowsTab';
+export { InfrastructureTab } from './InfrastructureTab';
+export { McpActivityTab } from './McpActivityTab';
+export { OverviewStrip } from './OverviewStrip';
