+++ packages/web/src/pages/Settings.tsx
@@ -1,12 +1,22 @@
 import { useState, useEffect, useCallback } from 'react';
-import { RefreshCw, Plus, Trash2, AlertCircle, CheckCircle, Mail, Info, Shield, Cpu, Link2, BookOpen, Mic, Send } from 'lucide-react';
+import { RefreshCw } from 'lucide-react';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Badge } from '@/components/ui/badge';
 import { Separator } from '@/components/ui/separator';
-import { skillsApi, triggersApi, adminApi, settingsApi, configApi, voiceSessionApi, wikiApi } from '@/lib/api';
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
+import { skillsApi, triggersApi, settingsApi, configApi, voiceSessionApi, wikiApi } from '@/lib/api';
 import type { Skill, Trigger, AutonomyLevel, AIRoutingResponse, IntegrationStatus, WikiLintReport } from '@/lib/types';
+import {
+  VersionUptimeSection,
+  ServiceHealthSection,
+  TriggersSection,
+  AutonomyLevelSection,
+  EmailAllowlistSection,
+  AIRoutingSection,
+  VoiceSection,
+  WikiSection,
+  EmailConfigSection,
+  IntegrationsSection,
+  DangerZoneSection,
+} from '@/components/settings';
 
 // ─── Types ────────────────────────────────────────────────────────────────────
 
@@ -23,1035 +33,6 @@ async function apiFetch<T>(path: string): Promise<T> {
   const res = await fetch(`${BASE}${path}`);
   if (!res.ok) throw new Error(`API ${res.status}`);
   return res.json() as Promise<T>;
-}
-
-function formatUptime(seconds?: number): string {
-  if (!seconds) return '—';
-  const h = Math.floor(seconds / 3600);
-  const m = Math.floor((seconds % 3600) / 60);
-  if (h > 0) return `${h}h ${m}m`;
-  return `${m}m`;
-}
-
-// ─── Shared ──────────────────────────────────────────────────────────────────
-
-function StatusDot({ status }: { status: 'up' | 'down' | 'degraded' | 'healthy' | 'unhealthy' | undefined }) {
-  if (status === 'up' || status === 'healthy') return <span className="inline-block w-2 h-2 rounded-full bg-green-500" />;
-  if (status === 'degraded') return <span className="inline-block w-2 h-2 rounded-full bg-yellow-500" />;
-  return <span className="inline-block w-2 h-2 rounded-full bg-red-500" />;
-}
-
-// ─── Version & Uptime section ────────────────────────────────────────────────
-
-function VersionUptimeSection({ version, uptime_s, loading, error }: {
-  version?: string;
-  uptime_s?: number;
-  loading: boolean;
-  error: string | null;
-}) {
-  return (
-    <section className="space-y-3">
-      <h2 className="text-base font-semibold">Version & Uptime</h2>
-
-      {error && (
-        <div className="flex items-center gap-2 rounded-lg border border-destructive/30 bg-destructive/10 px-3 py-2 text-sm text-destructive">
-          <AlertCircle className="h-4 w-4 shrink-0" />
-          {error}
-        </div>
-      )}
-
-      {loading && version === undefined && (
-        <div className="space-y-2">
-          {[...Array(2)].map((_, i) => <div key={i} className="h-8 animate-pulse rounded bg-secondary" />)}
-        </div>
-      )}
-
-      {version !== undefined && (
-        <div className="rounded-lg border bg-card divide-y">
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Version</span>
-            <span className="font-mono">{version ?? '—'}</span>
-          </div>
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Uptime</span>
-            <span>{formatUptime(uptime_s)}</span>
-          </div>
-        </div>
-      )}
-    </section>
-  );
-}
-
-// ─── Service Health section ──────────────────────────────────────────────────
-
-function ServiceHealthSection({ services, loading }: {
-  services?: Record<string, { status: 'up' | 'down' | 'degraded'; latency_ms?: number; models_available?: string[] }>;
-  loading: boolean;
-}) {
-  return (
-    <section className="space-y-3">
-      <h2 className="text-base font-semibold">Service Health</h2>
-
-      {loading && !services && (
-        <div className="space-y-2">
-          {[...Array(3)].map((_, i) => <div key={i} className="h-8 animate-pulse rounded bg-secondary" />)}
-        </div>
-      )}
-
-      {services && Object.keys(services).length > 0 ? (
-        <div className="rounded-lg border bg-card divide-y">
-          {Object.entries(services).map(([name, svc]) => (
-            <div key={name} className="px-4 py-3 flex items-center justify-between text-sm">
-              <div className="flex items-center gap-2">
-                <StatusDot status={svc.status} />
-                <span className="capitalize">{name === 'llm' ? 'LLM' : name}</span>
-              </div>
-              <div className="flex items-center gap-3 text-muted-foreground">
-                {svc.latency_ms !== undefined && <span>{svc.latency_ms}ms</span>}
-                {svc.models_available && svc.models_available.length > 0 && (
-                  <span className="text-xs">{svc.models_available.slice(0, 3).join(', ')}{svc.models_available.length > 3 ? '…' : ''}</span>
-                )}
-              </div>
-            </div>
-          ))}
-        </div>
-      ) : !loading && (
-        <p className="text-sm text-muted-foreground">No service data available.</p>
-      )}
-    </section>
-  );
-}
-
-
-
-
-
-/**
- * Convert a cron expression to a simple human-readable description.
- * Covers common patterns without pulling in a full library like cronstrue.
- */
-function describeCron(expr: string): string | null {
-  const trimmed = expr.trim();
-  const fields = trimmed.split(/\s+/);
-  if (fields.length !== 5) return null;
-
-  const [minute, hour, dom, month, dow] = fields;
-  const dayNames = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
-
-  // Every hour: "0 * * * *"
-  if (minute.match(/^\d+$/) && hour === '*' && dom === '*' && month === '*' && dow === '*') {
-    return `Every hour at minute ${minute}`;
-  }
-
-  // Daily at specific time: "M H * * *"
-  if (minute.match(/^\d+$/) && hour.match(/^\d+$/) && dom === '*' && month === '*' && dow === '*') {
-    const h = parseInt(hour, 10);
-    const m = parseInt(minute, 10);
-    const ampm = h >= 12 ? 'PM' : 'AM';
-    const h12 = h === 0 ? 12 : h > 12 ? h - 12 : h;
-    return `Daily at ${h12}:${m.toString().padStart(2, '0')} ${ampm}`;
-  }
-
-  // Weekly: "M H * * D"
-  if (minute.match(/^\d+$/) && hour.match(/^\d+$/) && dom === '*' && month === '*' && dow.match(/^\d$/)) {
-    const h = parseInt(hour, 10);
-    const m = parseInt(minute, 10);
-    const d = parseInt(dow, 10);
-    const ampm = h >= 12 ? 'PM' : 'AM';
-    const h12 = h === 0 ? 12 : h > 12 ? h - 12 : h;
-    const dayName = dayNames[d] ?? `day ${d}`;
-    return `Every ${dayName} at ${h12}:${m.toString().padStart(2, '0')} ${ampm}`;
-  }
-
-  // Every N minutes: "*/N * * * *"
-  if (minute.startsWith('*/') && hour === '*' && dom === '*' && month === '*' && dow === '*') {
-    return `Every ${minute.slice(2)} minutes`;
-  }
-
-  // Every N hours: "0 */N * * *"
-  if (minute === '0' && hour.startsWith('*/') && dom === '*' && month === '*' && dow === '*') {
-    return `Every ${hour.slice(2)} hours`;
-  }
-
-  return null;
-}
-
-
-
-// ─── Triggers section ─────────────────────────────────────────────────────────
-
-function TriggersSection({ triggers, loading, error, onAdd, onDelete }: {
-  triggers: Trigger[];
-  loading: boolean;
-  error: string | null;
-  onAdd: (name: string, queryText: string) => Promise<void>;
-  onDelete: (id: string) => Promise<void>;
-}) {
-  const [adding, setAdding] = useState(false);
-  const [name, setName] = useState('');
-  const [queryText, setQueryText] = useState('');
-  const [addSubmitting, setAddSubmitting] = useState(false);
-  const [addError, setAddError] = useState<string | null>(null);
-  const [deleting, setDeleting] = useState<string | null>(null);
-
-  async function handleAdd(e: React.FormEvent) {
-    e.preventDefault();
-    if (!name.trim() || !queryText.trim()) return;
-    setAddSubmitting(true);
-    setAddError(null);
-    try {
-      await onAdd(name.trim(), queryText.trim());
-      setName('');
-      setQueryText('');
-      setAdding(false);
-    } catch (err) {
-      setAddError(err instanceof Error ? err.message : 'Failed to add trigger');
-    } finally {
-      setAddSubmitting(false);
-    }
-  }
-
-  async function handleDelete(id: string) {
-    setDeleting(id);
-    try {
-      await onDelete(id);
-    } finally {
-      setDeleting(null);
-    }
-  }
-
-  return (
-    <section className="space-y-3">
-      <div className="flex items-center justify-between">
-        <h2 className="text-base font-semibold">Semantic Triggers</h2>
-        <Button size="sm" variant="outline" onClick={() => setAdding((v) => !v)} className="gap-1.5">
-          <Plus className="h-3.5 w-3.5" />
-          Add
-        </Button>
-      </div>
-
-      {error && (
-        <div className="flex items-center gap-2 rounded-lg border border-destructive/30 bg-destructive/10 px-3 py-2 text-sm text-destructive">
-          <AlertCircle className="h-4 w-4 shrink-0" />
-          {error}
-        </div>
-      )}
-
-      {adding && (
-        <form onSubmit={handleAdd} className="rounded-lg border bg-card p-4 space-y-3">
-          <h3 className="text-sm font-medium">New Trigger</h3>
-          <Input
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            placeholder="Name (e.g. client-risk-mentions)"
-            required
-          />
-          <Input
-            value={queryText}
-            onChange={(e) => setQueryText(e.target.value)}
-            placeholder="Semantic query (e.g. client escalation risk blowup)"
-            required
-          />
-          {addError && <p className="text-xs text-destructive">{addError}</p>}
-          <div className="flex gap-2">
-            <Button type="submit" size="sm" disabled={addSubmitting || !name.trim() || !queryText.trim()}>
-              {addSubmitting ? 'Adding...' : 'Add Trigger'}
-            </Button>
-            <Button type="button" size="sm" variant="ghost" onClick={() => setAdding(false)}>Cancel</Button>
-          </div>
-        </form>
-      )}
-
-      {loading && triggers.length === 0 ? (
-        <div className="space-y-2">
-          {[...Array(2)].map((_, i) => <div key={i} className="h-14 animate-pulse rounded bg-secondary" />)}
-        </div>
-      ) : triggers.length === 0 ? (
-        <div className="rounded-lg border border-dashed p-8 text-center text-muted-foreground">
-          <p className="text-sm">No triggers configured.</p>
-          <p className="text-xs mt-1">Triggers fire a Pushover notification when a new capture matches a semantic query.</p>
-        </div>
-      ) : (
-        <div className="rounded-lg border bg-card divide-y">
-          {triggers.map((trigger) => (
-            <div key={trigger.id} className="px-4 py-3 flex items-start justify-between gap-3">
-              <div className="flex-1 min-w-0">
-                <div className="flex items-center gap-2">
-                  <span className="text-sm font-medium">{trigger.name}</span>
-                  <Badge variant={(trigger.is_active ?? trigger.enabled) ? 'default' : 'secondary'} className="text-xs">
-                    {(trigger.is_active ?? trigger.enabled) ? 'active' : 'inactive'}
-                  </Badge>
-                  {trigger.delivery_channel && (
-                    <Badge variant="outline" className="text-xs capitalize">{trigger.delivery_channel}</Badge>
-                  )}
-                </div>
-                {trigger.query_text && (
-                  <p className="text-xs text-muted-foreground mt-0.5 font-mono truncate">{trigger.query_text}</p>
-                )}
-                <div className="text-xs text-muted-foreground mt-0.5 space-x-2">
-                  {trigger.threshold !== undefined && <span>Threshold: {trigger.threshold}</span>}
-                  {trigger.cooldown_minutes !== undefined && <span>Cooldown: {trigger.cooldown_minutes}m</span>}
-                  {trigger.fire_count !== undefined && <span>Fired: {trigger.fire_count}x</span>}
-                  {trigger.last_fired_at && (
-                    <span>Last: {new Date(trigger.last_fired_at).toLocaleString()}</span>
-                  )}
-                </div>
-              </div>
-              <Button
-                size="sm"
-                variant="ghost"
-                className="text-muted-foreground hover:text-destructive shrink-0"
-                disabled={deleting === trigger.id}
-                onClick={() => handleDelete(trigger.id)}
-                aria-label={`Delete trigger ${trigger.name}`}
-              >
-                <Trash2 className="h-3.5 w-3.5" />
-              </Button>
-            </div>
-          ))}
-        </div>
-      )}
-    </section>
-  );
-}
-
-// ─── Autonomy Level section ─────────────────────────────────────────────────
-
-const AUTONOMY_LEVELS: { value: AutonomyLevel; label: string; description: string }[] = [
-  { value: 'observe', label: 'Observe', description: 'Notifications only \u2014 no autonomous actions' },
-  { value: 'assist', label: 'Assist', description: 'Draft + notify \u2014 human relays responses' },
-  { value: 'advise', label: 'Advise', description: 'Act + report \u2014 posts with clear bot attribution' },
-  { value: 'partner', label: 'Partner', description: 'Autonomous within guardrails' },
-];
-
-function AutonomyLevelSection({ level, loading, error, onChange }: {
-  level: AutonomyLevel;
-  loading: boolean;
-  error: string | null;
-  onChange: (level: AutonomyLevel) => Promise<void>;
-}) {
-  const [saving, setSaving] = useState(false);
-  const [saveResult, setSaveResult] = useState<{ success: boolean; message: string } | null>(null);
-
-  async function handleChange(newLevel: AutonomyLevel) {
-    if (newLevel === level) return;
-    setSaving(true);
-    setSaveResult(null);
-    try {
-      await onChange(newLevel);
-      setSaveResult({ success: true, message: 'Saved' });
-      setTimeout(() => setSaveResult(null), 3000);
-    } catch (err) {
-      setSaveResult({ success: false, message: err instanceof Error ? err.message : 'Failed to save' });
-    } finally {
-      setSaving(false);
-    }
-  }
-
-  return (
-    <section className="space-y-3">
-      <div className="flex items-center gap-2">
-        <Shield className="h-4 w-4 text-muted-foreground" />
-        <h2 className="text-base font-semibold">Autonomy Level</h2>
-        {saveResult && (
-          <Badge variant={saveResult.success ? 'default' : 'destructive'} className="text-xs">
-            {saveResult.success ? <CheckCircle className="h-3 w-3 mr-1" /> : <AlertCircle className="h-3 w-3 mr-1" />}
-            {saveResult.message}
-          </Badge>
-        )}
-      </div>
-
-      {error && (
-        <div className="flex items-center gap-2 rounded-lg border border-destructive/30 bg-destructive/10 px-3 py-2 text-sm text-destructive">
-          <AlertCircle className="h-4 w-4 shrink-0" />
-          {error}
-        </div>
-      )}
-
-      {loading ? (
-        <div className="space-y-2">
-          {[...Array(4)].map((_, i) => <div key={i} className="h-12 animate-pulse rounded bg-secondary" />)}
-        </div>
-      ) : (
-        <div className="rounded-lg border bg-card divide-y">
-          {AUTONOMY_LEVELS.map((opt) => (
-            <label
-              key={opt.value}
-              className={`px-4 py-3 flex items-start gap-3 cursor-pointer transition-colors hover:bg-accent/50 ${
-                saving ? 'opacity-60 pointer-events-none' : ''
-              }`}
-            >
-              <input
-                type="radio"
-                name="autonomy_level"
-                value={opt.value}
-                checked={level === opt.value}
-                onChange={() => handleChange(opt.value)}
-                disabled={saving}
-                className="mt-0.5 accent-primary"
-              />
-              <div className="flex-1 min-w-0">
-                <div className="flex items-center gap-2">
-                  <span className="text-sm font-medium">{opt.label}</span>
-                  {level === opt.value && (
-                    <Badge variant="default" className="text-xs">Active</Badge>
-                  )}
-                </div>
-                <p className="text-xs text-muted-foreground mt-0.5">{opt.description}</p>
-              </div>
-            </label>
-          ))}
-        </div>
-      )}
-    </section>
-  );
-}
-
-// ─── Email Allowlist section ─────────────────────────────────────────────────
-
-function EmailAllowlistSection({ entries, loading, error, onAdd, onRemove }: {
-  entries: string[];
-  loading: boolean;
-  error: string | null;
-  onAdd: (entry: string) => Promise<void>;
-  onRemove: (entry: string) => Promise<void>;
-}) {
-  const [adding, setAdding] = useState(false);
-  const [newEntry, setNewEntry] = useState('');
-  const [addSubmitting, setAddSubmitting] = useState(false);
-  const [addError, setAddError] = useState<string | null>(null);
-  const [removing, setRemoving] = useState<string | null>(null);
-
-  function validateEntry(value: string): string | null {
-    const trimmed = value.trim().toLowerCase();
-    if (!trimmed) return 'Entry cannot be empty';
-    if (trimmed.startsWith('@')) {
-      // Domain pattern: @example.com
-      if (!/^@[a-z0-9.-]+\.[a-z]{2,}$/.test(trimmed)) return 'Invalid domain format. Use @example.com';
-      return null;
-    }
-    // Email address
-    if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(trimmed)) return 'Invalid email address';
-    return null;
-  }
-
-  async function handleAdd(e: React.FormEvent) {
-    e.preventDefault();
-    const trimmed = newEntry.trim().toLowerCase();
-    const validationError = validateEntry(trimmed);
-    if (validationError) {
-      setAddError(validationError);
-      return;
-    }
-    if (entries.includes(trimmed)) {
-      setAddError('Already in the allowlist');
-      return;
-    }
-    setAddSubmitting(true);
-    setAddError(null);
-    try {
-      await onAdd(trimmed);
-      setNewEntry('');
-      setAdding(false);
-    } catch (err) {
-      setAddError(err instanceof Error ? err.message : 'Failed to add entry');
-    } finally {
-      setAddSubmitting(false);
-    }
-  }
-
-  async function handleRemove(entry: string) {
-    setRemoving(entry);
-    try {
-      await onRemove(entry);
-    } catch {
-      setAddError('Failed to remove entry — please try again');
-    } finally {
-      setRemoving(null);
-    }
-  }
-
-  return (
-    <section className="space-y-3">
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <h2 className="text-base font-semibold">Email Allowlist</h2>
-          <TooltipProvider>
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <Info className="h-4 w-4 text-muted-foreground cursor-help" />
-              </TooltipTrigger>
-              <TooltipContent side="right" className="max-w-xs">
-                <p>Controls which senders can create captures via <strong>brain@troy-davis.com</strong>.</p>
-                <p className="mt-1">Enter a full email address (e.g. <code className="text-xs bg-muted px-1 rounded">user@example.com</code>) for exact matching, or a domain starting with <code className="text-xs bg-muted px-1 rounded">@</code> (e.g. <code className="text-xs bg-muted px-1 rounded">@example.com</code>) to allow all addresses from that domain.</p>
-              </TooltipContent>
-            </Tooltip>
-          </TooltipProvider>
-        </div>
-        <Button size="sm" variant="outline" onClick={() => setAdding((v) => !v)} className="gap-1.5">
-          <Plus className="h-3.5 w-3.5" />
-          Add
-        </Button>
-      </div>
-
-      {error && (
-        <div className="flex items-center gap-2 rounded-lg border border-destructive/30 bg-destructive/10 px-3 py-2 text-sm text-destructive">
-          <AlertCircle className="h-4 w-4 shrink-0" />
-          {error}
-        </div>
-      )}
-
-      {adding && (
-        <form onSubmit={handleAdd} className="rounded-lg border bg-card p-4 space-y-3">
-          <h3 className="text-sm font-medium">Add Allowed Sender</h3>
-          <Input
-            value={newEntry}
-            onChange={(e) => { setNewEntry(e.target.value); setAddError(null); }}
-            placeholder="user@example.com or @example.com"
-            required
-            autoFocus
-          />
-          {addError && <p className="text-xs text-destructive">{addError}</p>}
-          <div className="flex gap-2">
-            <Button type="submit" size="sm" disabled={addSubmitting || !newEntry.trim()}>
-              {addSubmitting ? 'Adding...' : 'Add'}
-            </Button>
-            <Button type="button" size="sm" variant="ghost" onClick={() => { setAdding(false); setAddError(null); }}>Cancel</Button>
-          </div>
-        </form>
-      )}
-
-      {loading && entries.length === 0 ? (
-        <div className="space-y-2">
-          {[...Array(3)].map((_, i) => <div key={i} className="h-10 animate-pulse rounded bg-secondary" />)}
-        </div>
-      ) : entries.length === 0 ? (
-        <div className="rounded-lg border border-dashed p-8 text-center text-muted-foreground">
-          <p className="text-sm">No allowed senders configured.</p>
-          <p className="text-xs mt-1">All emails to brain@troy-davis.com will be rejected until you add at least one entry.</p>
-        </div>
-      ) : (
-        <div className="rounded-lg border bg-card divide-y">
-          {entries.map((entry) => (
-            <div key={entry} className="px-4 py-3 flex items-center justify-between gap-3">
-              <div className="flex items-center gap-2 min-w-0">
-                <Mail className="h-3.5 w-3.5 text-muted-foreground shrink-0" />
-                <span className="text-sm font-mono truncate">{entry}</span>
-                <Badge variant={entry.startsWith('@') ? 'secondary' : 'outline'} className="text-xs shrink-0">
-                  {entry.startsWith('@') ? 'domain' : 'exact'}
-                </Badge>
-              </div>
-              <Button
-                size="sm"
-                variant="ghost"
-                className="text-muted-foreground hover:text-destructive shrink-0"
-                disabled={removing === entry}
-                onClick={() => handleRemove(entry)}
-                aria-label={`Remove ${entry}`}
-              >
-                <Trash2 className="h-3.5 w-3.5" />
-              </Button>
-            </div>
-          ))}
-        </div>
-      )}
-    </section>
-  );
-}
-
-// ─── AI Routing section ──────────────────────────────────────────────────────
-
-function AIRoutingSection({ routing, loading, error }: {
-  routing: AIRoutingResponse | null;
-  loading: boolean;
-  error: string | null;
-}) {
-  if (loading && !routing) {
-    return (
-      <section className="space-y-3">
-        <div className="flex items-center gap-2">
-          <Cpu className="h-4 w-4 text-muted-foreground" />
-          <h2 className="text-base font-semibold">AI Routing</h2>
-        </div>
-        <div className="space-y-2">
-          {[...Array(4)].map((_, i) => <div key={i} className="h-8 animate-pulse rounded bg-secondary" />)}
-        </div>
-      </section>
-    );
-  }
-
-  const budgetPct = routing?.budget
-    ? Math.min(100, (routing.budget.month_total_usd / routing.budget.hard_limit_usd) * 100)
-    : 0;
-  const budgetColor = budgetPct > 80 ? 'bg-destructive' : budgetPct > 60 ? 'bg-yellow-500' : 'bg-green-500';
-
-  return (
-    <section className="space-y-3">
-      <div className="flex items-center gap-2">
-        <Cpu className="h-4 w-4 text-muted-foreground" />
-        <h2 className="text-base font-semibold">AI Routing</h2>
-      </div>
-
-      {error && (
-        <div className="flex items-center gap-2 rounded-lg border border-destructive/30 bg-destructive/10 px-3 py-2 text-sm text-destructive">
-          <AlertCircle className="h-4 w-4 shrink-0" />
-          {error}
-        </div>
-      )}
-
-      {routing && (
-        <>
-          {/* Model routing table */}
-          <div className="rounded-lg border bg-card overflow-hidden">
-            <div className="grid grid-cols-[1fr_1fr_auto_auto] gap-x-4 px-4 py-2 text-xs font-medium text-muted-foreground border-b bg-muted/30">
-              <span>Task</span>
-              <span>Model</span>
-              <span className="text-right">Client</span>
-              <span className="text-right">Calls</span>
-            </div>
-            {routing.models.map((entry) => (
-              <div
-                key={entry.task}
-                className="grid grid-cols-[1fr_1fr_auto_auto] gap-x-4 px-4 py-2.5 text-sm border-b last:border-b-0"
-              >
-                <span className="font-medium capitalize">{entry.task}</span>
-                <span className="font-mono text-xs text-muted-foreground truncate">{entry.model}</span>
-                <Badge variant={entry.client === 'anthropic' ? 'default' : 'secondary'} className="text-xs justify-self-end">
-                  {entry.client}
-                </Badge>
-                <span className="text-xs text-muted-foreground text-right tabular-nums">
-                  {entry.month_calls > 0 ? entry.month_calls.toLocaleString() : '—'}
-                </span>
-              </div>
-            ))}
-          </div>
-
-          {/* Monthly budget progress */}
-          <div className="rounded-lg border bg-card px-4 py-3 space-y-2">
-            <div className="flex items-center justify-between text-sm">
-              <span className="text-muted-foreground">Monthly Spend</span>
-              <span className="font-mono text-xs">
-                ${routing.budget.month_total_usd.toFixed(2)} / ${routing.budget.hard_limit_usd.toFixed(0)} budget
-              </span>
-            </div>
-            <div className="h-2 rounded-full bg-secondary overflow-hidden">
-              <div
-                className={`h-full rounded-full transition-all ${budgetColor}`}
-                style={{ width: `${budgetPct}%` }}
-              />
-            </div>
-            <div className="flex items-center justify-between text-xs text-muted-foreground">
-              <span>Soft limit: ${routing.budget.soft_limit_usd}</span>
-              <span>Hard limit: ${routing.budget.hard_limit_usd}</span>
-            </div>
-          </div>
-        </>
-      )}
-    </section>
-  );
-}
-
-// ─── Voice section ──────────────────────────────────────────────────────────
-
-function VoiceSection({ integrations, voiceStats, loading }: {
-  integrations: IntegrationStatus[];
-  voiceStats: { totalSessions: number; activeSessions: number } | null;
-  loading: boolean;
-}) {
-  const pipecatIntegration = integrations.find(i => i.name.toLowerCase().includes('pipecat') || i.name.toLowerCase().includes('voice'));
-  const voiceCaptureIntegration = integrations.find(i => i.name === 'Voice Capture');
-
-  return (
-    <section className="space-y-3">
-      <div className="flex items-center gap-2">
-        <Mic className="h-4 w-4 text-muted-foreground" />
-        <h2 className="text-base font-semibold">Voice</h2>
-      </div>
-
-      {loading && !voiceStats ? (
-        <div className="space-y-2">
-          {[...Array(3)].map((_, i) => <div key={i} className="h-8 animate-pulse rounded bg-secondary" />)}
-        </div>
-      ) : (
-        <div className="rounded-lg border bg-card divide-y">
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Pipecat (Conversational)</span>
-            {pipecatIntegration ? (
-              <div className="flex items-center gap-2">
-                <span className={`inline-block w-2 h-2 rounded-full ${pipecatIntegration.status === 'connected' ? 'bg-green-500' : 'bg-red-500'}`} />
-                <span className="text-xs text-muted-foreground">{pipecatIntegration.detail ?? pipecatIntegration.status}</span>
-              </div>
-            ) : (
-              <Badge variant="secondary" className="text-xs">Not configured</Badge>
-            )}
-          </div>
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Voice Capture (iOS Shortcut)</span>
-            {voiceCaptureIntegration ? (
-              <div className="flex items-center gap-2">
-                <span className={`inline-block w-2 h-2 rounded-full ${voiceCaptureIntegration.status === 'connected' ? 'bg-green-500' : 'bg-red-500'}`} />
-                <span className="text-xs text-muted-foreground">{voiceCaptureIntegration.detail ?? voiceCaptureIntegration.status}</span>
-              </div>
-            ) : (
-              <Badge variant="default" className="text-xs">Active</Badge>
-            )}
-          </div>
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Total Sessions</span>
-            <span className="font-mono text-xs">{voiceStats?.totalSessions ?? '—'}</span>
-          </div>
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Active Sessions</span>
-            <span className={`font-mono text-xs ${(voiceStats?.activeSessions ?? 0) > 0 ? 'text-blue-600 dark:text-blue-400' : ''}`}>
-              {voiceStats?.activeSessions ?? 0}
-            </span>
-          </div>
-        </div>
-      )}
-    </section>
-  );
-}
-
-// ─── Wiki section ────────────────────────────────────────────────────────────
-
-function WikiSection({ skills, integrations, wikiStats, loading }: {
-  skills: Skill[];
-  integrations: IntegrationStatus[];
-  wikiStats: { pageCount: number; lastLintRun: string | null; lastSync: string | null } | null;
-  loading: boolean;
-}) {
-  // Wiki-related skills for schedule display
-  const wikiLintSkill = skills.find(s => s.name === 'wiki-lint');
-  const wikiIngestSkill = skills.find(s => s.name === 'wiki-ingest' || s.name === 'wiki-synthesis');
-  const giteaIntegration = integrations.find(i => i.name === 'Gitea');
-  const giteaUrl = giteaIntegration?.url;
-  const giteaConfigured = giteaIntegration?.status === 'connected';
-
-  return (
-    <section className="space-y-3">
-      <div className="flex items-center gap-2">
-        <BookOpen className="h-4 w-4 text-muted-foreground" />
-        <h2 className="text-base font-semibold">Wiki</h2>
-      </div>
-
-      {loading ? (
-        <div className="space-y-2">
-          {[...Array(3)].map((_, i) => <div key={i} className="h-8 animate-pulse rounded bg-secondary" />)}
-        </div>
-      ) : (
-        <div className="rounded-lg border bg-card divide-y">
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Gitea Repo</span>
-            <div className="flex items-center gap-2">
-              <span className={`inline-block w-2 h-2 rounded-full ${giteaConfigured ? 'bg-green-500' : 'bg-red-500'}`} />
-              <span className="font-mono text-xs text-muted-foreground truncate max-w-[250px]">
-                {giteaUrl ?? 'Not configured'}
-              </span>
-            </div>
-          </div>
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Page Count</span>
-            <span className="font-mono text-xs">{wikiStats?.pageCount ?? '—'}</span>
-          </div>
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Last Sync</span>
-            <span className="text-xs text-muted-foreground">
-              {wikiStats?.lastSync ? new Date(wikiStats.lastSync).toLocaleString() : '—'}
-            </span>
-          </div>
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Lint Schedule</span>
-            <span className="font-mono text-xs">
-              {wikiLintSkill?.schedule
-                ? `${wikiLintSkill.schedule}${describeCron(wikiLintSkill.schedule) ? ` (${describeCron(wikiLintSkill.schedule)})` : ''}`
-                : '—'}
-            </span>
-          </div>
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Last Lint</span>
-            <span className="text-xs text-muted-foreground">
-              {wikiStats?.lastLintRun ? new Date(wikiStats.lastLintRun).toLocaleString() : '—'}
-            </span>
-          </div>
-          {wikiIngestSkill && (
-            <div className="px-4 py-3 flex items-center justify-between text-sm">
-              <span className="text-muted-foreground">Auto-ingest</span>
-              <Badge variant={wikiIngestSkill.schedule ? 'default' : 'secondary'} className="text-xs">
-                {wikiIngestSkill.schedule ? 'Enabled' : 'Disabled'}
-              </Badge>
-            </div>
-          )}
-        </div>
-      )}
-    </section>
-  );
-}
-
-// ─── Email Configuration section ────────────────────────────────────────────
-
-function EmailConfigSection({ integrations, loading }: {
-  integrations: IntegrationStatus[];
-  loading: boolean;
-}) {
-  const inbound = integrations.find(i => i.name === 'Email (Inbound)');
-  const outbound = integrations.find(i => i.name === 'Email (Outbound)');
-
-  return (
-    <section className="space-y-3">
-      <div className="flex items-center gap-2">
-        <Send className="h-4 w-4 text-muted-foreground" />
-        <h2 className="text-base font-semibold">Email Configuration</h2>
-      </div>
-
-      {loading && !inbound && !outbound ? (
-        <div className="space-y-2">
-          {[...Array(2)].map((_, i) => <div key={i} className="h-8 animate-pulse rounded bg-secondary" />)}
-        </div>
-      ) : (
-        <div className="rounded-lg border bg-card divide-y">
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Inbound</span>
-            <div className="flex items-center gap-2">
-              <span className={`inline-block w-2 h-2 rounded-full ${inbound?.status === 'connected' ? 'bg-green-500' : 'bg-red-500'}`} />
-              <span className="text-xs text-muted-foreground">
-                {inbound?.detail ?? 'Not configured'}
-              </span>
-            </div>
-          </div>
-          <div className="px-4 py-3 flex items-center justify-between text-sm">
-            <span className="text-muted-foreground">Outbound</span>
-            <div className="flex items-center gap-2">
-              <span className={`inline-block w-2 h-2 rounded-full ${outbound?.status === 'connected' ? 'bg-green-500' : 'bg-red-500'}`} />
-              <span className="text-xs text-muted-foreground">
-                {outbound?.detail ?? 'Not configured'}
-              </span>
-            </div>
-          </div>
-        </div>
-      )}
-    </section>
-  );
-}
-
-// ─── Integrations section ────────────────────────────────────────────────────
-
-function IntegrationsSection({ integrations, loading, error }: {
-  integrations: IntegrationStatus[];
-  loading: boolean;
-  error: string | null;
-}) {
-  return (
-    <section className="space-y-3">
-      <div className="flex items-center gap-2">
-        <Link2 className="h-4 w-4 text-muted-foreground" />
-        <h2 className="text-base font-semibold">Integrations</h2>
-      </div>
-
-      {error && (
-        <div className="flex items-center gap-2 rounded-lg border border-destructive/30 bg-destructive/10 px-3 py-2 text-sm text-destructive">
-          <AlertCircle className="h-4 w-4 shrink-0" />
-          {error}
-        </div>
-      )}
-
-      {loading && integrations.length === 0 ? (
-        <div className="space-y-2">
-          {[...Array(5)].map((_, i) => <div key={i} className="h-12 animate-pulse rounded bg-secondary" />)}
-        </div>
-      ) : integrations.length === 0 ? (
-        <p className="text-sm text-muted-foreground">No integration data available.</p>
-      ) : (
-        <div className="rounded-lg border bg-card divide-y">
-          {integrations.map((integration) => (
-            <div key={integration.name} className="px-4 py-3 flex items-center justify-between gap-3">
-              <div className="flex items-center gap-3 min-w-0">
-                <span className={`inline-block w-2 h-2 rounded-full shrink-0 ${
-                  integration.status === 'connected' ? 'bg-green-500' : 'bg-red-500'
-                }`} />
-                <div className="min-w-0">
-                  <span className="text-sm font-medium">{integration.name}</span>
-                  {integration.detail && (
-                    <p className="text-xs text-muted-foreground truncate">{integration.detail}</p>
-                  )}
-                </div>
-              </div>
-              <div className="flex items-center gap-2 shrink-0 text-right">
-                {integration.url && (
-                  <span className="text-xs text-muted-foreground font-mono hidden sm:inline truncate max-w-[200px]">
-                    {integration.url}
-                  </span>
-                )}
-                {integration.last_activity && (
-                  <TooltipProvider>
-                    <Tooltip>
-                      <TooltipTrigger asChild>
-                        <span className="text-xs text-muted-foreground cursor-help">
-                          Last: {new Date(integration.last_activity).toLocaleDateString()}
-                        </span>
-                      </TooltipTrigger>
-                      <TooltipContent>
-                        {new Date(integration.last_activity).toLocaleString()}
-                      </TooltipContent>
-                    </Tooltip>
-                  </TooltipProvider>
-                )}
-                <Badge
-                  variant={integration.status === 'connected' ? 'default' : 'destructive'}
-                  className="text-xs"
-                >
-                  {integration.status}
-                </Badge>
-              </div>
-            </div>
-          ))}
-        </div>
-      )}
-    </section>
-  );
-}
-
-// ─── Danger Zone section ──────────────────────────────────────────────────────
-
-const CONFIRM_PHRASE = 'WIPE ALL DATA';
-
-function DangerZoneSection() {
-  const [open, setOpen] = useState(false);
-  const [confirmText, setConfirmText] = useState('');
-  const [wiping, setWiping] = useState(false);
-  const [result, setResult] = useState<{ success: boolean; message: string } | null>(null);
-
-  const confirmed = confirmText === CONFIRM_PHRASE;
-
-  function handleOpen() {
-    setOpen(true);
-    setConfirmText('');
-    setResult(null);
-  }
-
-  function handleClose() {
-    setOpen(false);
-    setConfirmText('');
-  }
-
-  async function handleWipe() {
-    setWiping(true);
-    setResult(null);
-    try {
-      const res = await adminApi.resetData();
-      setResult({
-        success: true,
-        message: `Wiped ${res.cleared.length} tables. The brain is empty — ready for real data.`,
-      });
-      handleClose();
-    } catch (err) {
-      setResult({ success: false, message: err instanceof Error ? err.message : 'Wipe failed' });
-    } finally {
-      setWiping(false);
-    }
-  }
-
-  return (
-    <>
-      <section className="space-y-3">
-        <h2 className="text-base font-semibold text-destructive">Danger Zone</h2>
-        <div className="rounded-lg border border-destructive/30 bg-destructive/5 p-4 space-y-3">
-          <div className="flex items-start justify-between gap-4">
-            <div>
-              <p className="text-sm font-medium">Wipe All Data</p>
-              <p className="text-xs text-muted-foreground mt-0.5">
-                Permanently deletes all captures, entities, sessions, briefs, bets, and AI audit logs.
-                Schema, migration history, and semantic triggers are preserved. Cannot be undone.
-              </p>
-            </div>
-            <Button
-              variant="destructive"
-              size="sm"
-              onClick={handleOpen}
-              className="shrink-0 gap-1.5"
-            >
-              <Trash2 className="h-3.5 w-3.5" />
-              Wipe All Data
-            </Button>
-          </div>
-          {result && (
-            <div className={`flex items-center gap-2 rounded px-3 py-2 text-sm border ${
-              result.success
-                ? 'bg-green-50 text-green-800 border-green-200'
-                : 'bg-destructive/10 text-destructive border-destructive/30'
-            }`}>
-              {result.success
-                ? <CheckCircle className="h-4 w-4 shrink-0" />
-                : <AlertCircle className="h-4 w-4 shrink-0" />}
-              {result.message}
-            </div>
-          )}
-        </div>
-      </section>
-
-      {/* Confirmation modal — rendered at root level to avoid layout clipping */}
-      {open && (
-        <div className="fixed inset-0 z-50 flex items-center justify-center p-4 bg-background/80 backdrop-blur-sm">
-          <div className="bg-card border rounded-lg shadow-lg w-full max-w-md p-6 space-y-4">
-            <div className="flex items-center gap-3">
-              <div className="rounded-full bg-destructive/10 p-2 shrink-0">
-                <AlertCircle className="h-5 w-5 text-destructive" />
-              </div>
-              <h3 className="text-base font-semibold">Wipe All Data?</h3>
-            </div>
-
-            <p className="text-sm text-muted-foreground">The following will be <strong>permanently deleted</strong>:</p>
-            <ul className="text-sm text-muted-foreground space-y-1 list-disc list-inside">
-              <li>All captures and embeddings</li>
-              <li>All entities and relationships</li>
-              <li>All governance sessions and messages</li>
-              <li>All weekly briefs (skills log)</li>
-              <li>All AI audit logs and bets</li>
-            </ul>
-            <p className="text-sm text-muted-foreground">
-              <strong>Preserved:</strong> semantic triggers, schema, migration history.
-            </p>
-
-            <div className="space-y-1.5">
-              <p className="text-xs font-medium">
-                Type{' '}
-                <span className="font-mono bg-muted px-1.5 py-0.5 rounded text-foreground">{CONFIRM_PHRASE}</span>
-                {' '}to confirm:
-              </p>
-              <Input
-                value={confirmText}
-                onChange={(e) => setConfirmText(e.target.value)}
-                placeholder={CONFIRM_PHRASE}
-                className="font-mono"
-                autoFocus
-                onKeyDown={(e) => e.key === 'Escape' && handleClose()}
-              />
-            </div>
-
-            {result && !result.success && (
-              <div className="flex items-center gap-2 rounded px-3 py-2 text-sm border bg-destructive/10 text-destructive border-destructive/30">
-                <AlertCircle className="h-4 w-4 shrink-0" />
-                {result.message}
-              </div>
-            )}
-
-            <div className="flex gap-2 justify-end pt-1">
-              <Button variant="ghost" size="sm" onClick={handleClose} disabled={wiping}>
-                Cancel
-              </Button>
-              <Button
-                variant="destructive"
-                size="sm"
-                disabled={!confirmed || wiping}
-                onClick={handleWipe}
-              >
-                {wiping ? 'Wiping...' : 'Confirm Wipe'}
-              </Button>
-            </div>
-          </div>
-        </div>
-      )}
-    </>
-  );
 }
 
 // ─── Main page ────────────────────────────────────────────────────────────────
@@ -1091,6 +72,8 @@ export default function Settings() {
   const [wikiStatsLoading, setWikiStatsLoading] = useState(true);
 
   const [refreshing, setRefreshing] = useState(false);
+
+  // ─── Data loaders ───────────────────────────────────────────────────────────
 
   const loadHealth = useCallback(async () => {
     setHealthError(null);
@@ -1133,7 +116,6 @@ export default function Settings() {
       const res = await settingsApi.get<string>('autonomy_level');
       setAutonomyLevel(res.value as AutonomyLevel);
     } catch (err) {
-      // 404 means no level set yet — default to observe
       if (err instanceof Error && err.message.includes('404')) {
         setAutonomyLevel('observe');
       } else {
@@ -1150,7 +132,6 @@ export default function Settings() {
       const res = await settingsApi.get<string[]>('email_allowlist');
       setAllowlist(res.value);
     } catch (err) {
-      // 404 means no allowlist yet — treat as empty
       if (err instanceof Error && err.message.includes('404')) {
         setAllowlist([]);
       } else {
@@ -1167,7 +148,6 @@ export default function Settings() {
       const res = await configApi.aiRouting();
       setAiRouting(res);
     } catch (err) {
-      // Not critical — may not be deployed yet
       if (err instanceof Error && err.message.includes('404')) {
         setAiRouting(null);
       } else {
@@ -1205,7 +185,6 @@ export default function Settings() {
         activeSessions: activeRes.status === 'fulfilled' ? activeRes.value.sessions.length : 0,
       });
     } catch {
-      // Voice sessions may not be deployed — show zeros gracefully
       setVoiceStats({ totalSessions: 0, activeSessions: 0 });
     } finally {
       setVoiceStatsLoading(false);
@@ -1227,7 +206,6 @@ export default function Settings() {
           : null,
       });
     } catch {
-      // Wiki may not be deployed — show graceful fallback
       setWikiStats({ pageCount: 0, lastLintRun: null, lastSync: null });
     } finally {
       setWikiStatsLoading(false);
@@ -1241,6 +219,8 @@ export default function Settings() {
   useEffect(() => {
     loadAll();
   }, [loadAll]);
+
+  // ─── Handlers ───────────────────────────────────────────────────────────────
 
   async function handleRefresh() {
     setRefreshing(true);
@@ -1275,9 +255,10 @@ export default function Settings() {
     setAllowlist(updated);
   }
 
+  // ─── Render ─────────────────────────────────────────────────────────────────
+
   return (
     <div className="space-y-8">
-      {/* Header */}
       <div className="flex items-center justify-between">
         <h1 className="text-2xl font-bold">Settings</h1>
         <Button variant="outline" size="sm" onClick={handleRefresh} disabled={refreshing} className="gap-2">
@@ -1286,91 +267,26 @@ export default function Settings() {
         </Button>
       </div>
 
-      <VersionUptimeSection
-        version={health?.version}
-        uptime_s={health?.uptime_s}
-        loading={healthLoading}
-        error={healthError}
-      />
-
+      <VersionUptimeSection version={health?.version} uptime_s={health?.uptime_s} loading={healthLoading} error={healthError} />
       <Separator />
-
-      <ServiceHealthSection
-        services={health?.services}
-        loading={healthLoading}
-      />
-
+      <ServiceHealthSection services={health?.services} loading={healthLoading} />
       <Separator />
-
-      <TriggersSection
-        triggers={triggers}
-        loading={triggersLoading}
-        error={triggersError}
-        onAdd={handleAddTrigger}
-        onDelete={handleDeleteTrigger}
-      />
-
+      <TriggersSection triggers={triggers} loading={triggersLoading} error={triggersError} onAdd={handleAddTrigger} onDelete={handleDeleteTrigger} />
       <Separator />
-
-      <AutonomyLevelSection
-        level={autonomyLevel}
-        loading={autonomyLoading}
-        error={autonomyError}
-        onChange={handleAutonomyChange}
-      />
-
+      <AutonomyLevelSection level={autonomyLevel} loading={autonomyLoading} error={autonomyError} onChange={handleAutonomyChange} />
       <Separator />
-
-      <EmailAllowlistSection
-        entries={allowlist}
-        loading={allowlistLoading}
-        error={allowlistError}
-        onAdd={handleAddAllowlistEntry}
-        onRemove={handleRemoveAllowlistEntry}
-      />
-
+      <EmailAllowlistSection entries={allowlist} loading={allowlistLoading} error={allowlistError} onAdd={handleAddAllowlistEntry} onRemove={handleRemoveAllowlistEntry} />
       <Separator />
-
-      <AIRoutingSection
-        routing={aiRouting}
-        loading={aiRoutingLoading}
-        error={aiRoutingError}
-      />
-
+      <AIRoutingSection routing={aiRouting} loading={aiRoutingLoading} error={aiRoutingError} />
       <Separator />
-
-      <VoiceSection
-        integrations={integrations}
-        voiceStats={voiceStats}
-        loading={voiceStatsLoading}
-      />
-
+      <VoiceSection integrations={integrations} voiceStats={voiceStats} loading={voiceStatsLoading} />
       <Separator />
-
-      <WikiSection
-        skills={skills}
-        integrations={integrations}
-        wikiStats={wikiStats}
-        loading={skillsLoading || wikiStatsLoading}
-      />
-
+      <WikiSection skills={skills} integrations={integrations} wikiStats={wikiStats} loading={skillsLoading || wikiStatsLoading} />
       <Separator />
-
-      <EmailConfigSection
-        integrations={integrations}
-        loading={integrationsLoading}
-      />
-
+      <EmailConfigSection integrations={integrations} loading={integrationsLoading} />
       <Separator />
-
-      <IntegrationsSection
-        integrations={integrations}
-        loading={integrationsLoading}
-        error={integrationsError}
-      />
-
+      <IntegrationsSection integrations={integrations} loading={integrationsLoading} error={integrationsError} />
       <Separator />
-
       <DangerZoneSection />
     </div>
   );
